requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deeppyer==1.0.1
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
llvmlite==0.43.0
lumaai==1.5.0
multidict==6.1.0
nltk==3.9.1
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
propcache==0.2.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sqlitedict==2.1.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
yarl==1.17.1
